Use set for paths-to-update comparison
With about 140k tracks the update operation takes a long time. A
flamegraph shows update_library_convert_extra_info as taking almost all
of that time and slice_contains in particular.

With the previous release 0.2.9 updates where very fast and got to the
actual analyzing part right away. With 0.3.2 it spends a lot of time
before it even gets to analyzing. And it seems to be slower to start up
the more songs you have analyzed.

Blissify 0.2.9 seemed to use a HashSet too :)