Enable playlists based on multiple songs
This commit introduces functionality for generating playlist based on a
set of songs.

For good performance, I also introduce a new distance measure, extended
isolation distance.

While the previous distance metrics, euclidean distance and cosine
distance, could be made to measure distance to a set of songs, the
performance will not be as good.