feat(symphonia): implement "correct" conversion from stereo to mono
I was able to figure out that I could recover the behavior of ffmpeg by multiplying the average by sqrt(2), this apparently has to do with how ffmpeg corrects the volume (gain?) when doing the conversion.
I'm not sure how it scales to 2.1/5.1 surround, so for those we still do a naive average.

see: https://sound.stackexchange.com/questions/22516/how-to-do-mono-to-stereo-conversion