feat(symphonia): switch resampling to FftFixedIn instead of FastFixedIn
This method resamples the input signal in chunks using the FFT. It is more accurate (as measured by the difference between the output of the ffmpeg and symphonia decoders) while being about as fast as the previous method.

I also feature gated all the tests so they only run when valid codecs are available.