feat: implement symphonia-based decoder (#96)
* feat: implement symphonia-based decoder

upstreamed from `mecomp-analysis`

TODO: extract metadata as well.

* refactor(symphonia): revert to FastFixedIn for resampling for better performance

cuts the time it takes to run all the tests in half, on my machine

* refactor: remove SymphoniaDecoderError variant from BlissError enum

* ci: update CI to cover symphonia decoder

* chore: remove symphonia from default features

* feat(bench): add benchmark for symphonia

* feat(symphonia): implement "correct" conversion from stereo to mono

I was able to figure out that I could recover the behavior of ffmpeg by multiplying the average by sqrt(2), this apparently has to do with how ffmpeg corrects the volume (gain?) when doing the conversion.
I'm not sure how it scales to 2.1/5.1 surround, so for those we still do a naive average.

see: https://sound.stackexchange.com/questions/22516/how-to-do-mono-to-stereo-conversion

* test: add analysis test for symphonia

* test: add resampling test for mono audio file

* feat(symphonia): switch resampling to FftFixedIn instead of FastFixedIn

This method resamples the input signal in chunks using the FFT. It is more accurate (as measured by the difference between the output of the ffmpeg and symphonia decoders) while being about as fast as the previous method.

I also feature gated all the tests so they only run when valid codecs are available.

* test(symphonia): add comparison test between Symphonia and FFmpeg for various audio files

* test(benchmarks): add additonal benchmarks for decoding

* misc: clippy and typo fix

* feat(symphonia): improve downmixing (stereo->mono) performance

Improve the performance of the stereo->mono conversion, and implement size_hint for the symphonia decoder

* refactor(symphonia): remove some unecessary allocations and copies

* feat(symphonia): refactor decoding and resampling logic for improved performance and clarity

* test(benchmarks): add benchmark for MP3 decoding

* refactor(symphonia): extract the "audio decoding" part of the decoder into a separate struct, so it can be reused by other symphonia-based decoders

* test(decoder): long song benchmark

* refactor(song/test): consolidate expected analysis into a constant for reuse by other decoders